+++ CHANGELOG.md
@@ -40,6 +40,10 @@
 * IntuneDeviceConfigurationSharedMultiDevicePolicyWindows10
   * Add missing `AccessTokens` parameter to `Export-TargetResource`
     FIXES [#5034](https://github.com/microsoft/Microsoft365DSC/issues/5034)
+* IntuneSettingCatalogCustomPolicyWindows10
+  * Update export logic to target more specific policy types.
+  * Prevent thrown exception to be caught by exception handler.
+    FIXES [#5088](https://github.com/microsoft/Microsoft365DSC/issues/5088)
 * M365DSCUtil
   * Fixes an issue where the O365Group workload was not properly detected.
     FIXES [#5095](https://github.com/microsoft/Microsoft365DSC/issues/5095)